requirements.txt
hyperliquid-python-sdk
python-dotenv
eth-account
aiohttp
uvloop

